Remove redondant cast
Former-commit-id: d3715faeb524ef77a75cd850e18be8e6341d8b25 [formerly 09d11c3ebbd27d61c66b5486aca41efd107ba8b8]
Former-commit-id: 5a6e9a44de0fa1c5525907586257e5f8233710a1 [formerly b5e8d02cfc5165f938b37ede021606e5bdf444aa]
Former-commit-id: 67d8f7c2015e47816b7db43ff5fbd1e9cc9579cb
Former-commit-id: 2ec2386967389c5b74c3d838d956b7e6e80f45e8

+++ Source/Raw.Net/Source/Base/Common.cs
@@ -14,7 +14,7 @@ namespace RawNet
         {
             int _y_temp = x >> n;
             if ((_y_temp != 0))
-                x = (int)~_y_temp >> (32 - n);
+                x = ~_y_temp >> (32 - n);
             return x;
         }
